% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micerf-package.R
\docType{package}
\name{micerf-package}
\alias{micerf}
\alias{micerf-package}
\title{micerf: Fast, Lightweight MICE using Random Forests}
\description{
{micerf} is a stripped-down version of Multiple Imputation using Chained Equations (MICE) with Fully Conditional Specification (FCS) of equations and either random forests or sampling from training values as models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jesse-smith/micerf}
  \item \url{https://jesse-smith.github.io/micerf/}
  \item Report bugs at \url{https://github.com/jesse-smith/micerf/issues}
}

}
\author{
\strong{Maintainer}: Jesse Smith \email{jesse.smith@stjude.org} (\href{https://orcid.org/0000-0001-7375-2463}{ORCID})

}
\keyword{internal}
