% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mids.R
\name{Mids}
\alias{Mids}
\title{\code{{R6}} Class Containing the State of a MICE Model + Data Set}
\description{
\code{{R6}} Class Containing the State of a MICE Model + Data Set

\code{{R6}} Class Containing the State of a MICE Model + Data Set
}
\details{
\code{Mids} is analogous to the \code{{mice}} packages \code{S3} \code{mids} object, but ensures
that its contents are consistent and correct to the extent possible.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{\verb{[data.table]} The data to impute. Can only be modified
(by reference or assignment) while \code{self$iteration == 0}.}

\item{\code{m}}{\verb{[integer(1)]} The number of chains (imputed data sets)
to create. Can only be modified while \code{self$iteration == 0}.}

\item{\code{predictor_matrix}}{\verb{[logical matrix]} A matrix with rows
representing response variables (to impute) and columns representing
predictor variables (to use in imputation). If assigning,
must be a square matrix with the same number of rows and columns as
variables in \code{self$data}, and row/column ordering must match column
order in \code{self$data}.}

\item{\code{where}}{\verb{[logical matrix]} A matrix of the same dimensions as
\code{self$data} indicating what values to impute (\code{TRUE}) or ignore
(\code{FALSE}). Can only be modified while \code{self$iteration == 0}.}

\item{\code{iteration}}{\verb{[integer(1)]} The number of iterations performed. Can
only be assigned a value greater than or equal to the current value.}

\item{\code{chunk_size}}{\verb{[integer(1)]} The average number of chains to assign
to a \code{future}. Can only be modified while \code{self$iteration == 0}.}

\item{\code{seed}}{\verb{[integer(1)]} Seed for random number generation.
Can be \code{NULL}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Mids-new}{\code{Mids$new()}}
\item \href{#method-Mids-clone}{\code{Mids$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mids-new"></a>}}
\if{latex}{\out{\hypertarget{method-Mids-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{Mids} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mids$new(
  data,
  m = 5L,
  predictor_matrix = make_predictor_matrix(data),
  where = make_where(data),
  chunk_size = 1L,
  seed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Mids} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mids-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Mids-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mids$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
